requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
click==8.1.3
Flask==1.1.2
flask-restplus==0.13.0
gunicorn==20.0.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==4.17.3
MarkupSafe==1.1.1
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
rq==1.11.1
six==1.16.0
SQLAlchemy==1.4.54
Werkzeug==0.16.1
zipp==3.12.0
